+++ adapters/Maio/MaioAdapter/GADMAdapterMaioAdsManager.m
@@ -97,19 +97,19 @@ static NSMutableDictionary<NSString *, GADMAdapterMaioAdsManager *> *instances;
 
 - (NSError *)loadAdForZoneId:(NSString *)zoneId delegate:(id<MaioDelegate>)delegate {
   if ([self getAdapterForZoneID:zoneId]) {
-    NSString *errorDesc =
-        [NSString stringWithFormat:@"Maio does not supporting requesting a second ad for the same "
-                                   @"zone ID while the first request is still in progress."];
-    NSDictionary *errorInfo =
-        [NSDictionary dictionaryWithObjectsAndKeys:errorDesc, NSLocalizedDescriptionKey, nil];
-    return [NSError errorWithDomain:kGADMMaioErrorDomain code:0 userInfo:errorInfo];
-  } else {
-    [self addAdapter:delegate forZoneID:zoneId];
+    return [GADMMaioError
+        errorWithDescription:@"Maio does not supporting requesting a second ad for the same zone "
+                             @"ID while the first request is still in progress."];
   }
 
-  if ([self.maioInstance canShowAtZoneId:zoneId]) {
-    [delegate maioDidChangeCanShow:zoneId newValue:YES];
+  // If maio does not have an ad ready to be shown, then we fail the ad request to avoid timeouts.
+  if (![self.maioInstance canShowAtZoneId:zoneId]) {
+    return GADMAdapterMaioErrorWithCodeAndDescription(GADMAdapterMaioErrorAdNotAvailable,
+                                                      @"No ad available.");
   }
+
+  [self addAdapter:delegate forZoneID:zoneId];
+  [delegate maioDidChangeCanShow:zoneId newValue:YES];
   return nil;
 }