+++ adapters/Maio/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## maio iOS Mediation Adapter Changelog
 
+#### Next Version
+- To limit timeouts, ad requests will now immediately fail if maio does not have an ad ready to be shown.
+
 #### Version 1.5.3.0
 - Verified compatibility with maio SDK version 1.5.3.